DEV: move preferences files to lib
