+++ core/src/js/models/preferences/ftue.ts
@@ -1,6 +1,1 @@
-export class Ftue {
-	readonly hasSeenGlobalFtue: boolean = false;
-	readonly hasSeenAchievementsLoginButton: boolean = false;
-}
-
-export type FtueKey = undefined | 'hasSeenAchievementsLoginButton' | 'hasSeenGlobalFtue';
+export { Ftue, FtueKey } from '@firestone/shared/feature/preferences';
