+++ libs/shared/feature/preferences/src/index.ts
@@ -1,1 +1,4 @@
+export * from './lib/ftue.model';
+export * from './lib/preferences.model';
+export * from './lib/preferences.service';
 export * from './lib/shared-feature-preferences.module';
