+++ libs/shared/feature/preferences/src/lib/ftue.model.ts
@@ -1,0 +1,6 @@
+export class Ftue {
+	readonly hasSeenGlobalFtue: boolean = false;
+	readonly hasSeenAchievementsLoginButton: boolean = false;
+}
+
+export type FtueKey = undefined | 'hasSeenAchievementsLoginButton' | 'hasSeenGlobalFtue';
