+++ libs/shared/feature/preferences/src/lib/shared-feature-preferences.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { PreferencesService } from './preferences.service';
 
 @NgModule({
 	imports: [CommonModule],
+	providers: [PreferencesService],
 })
 export class SharedFeaturePreferencesModule {}
